create rights for users to delete their dogs or posts

+++ app/controllers/dogs_controller.rb
@@ -11,6 +11,9 @@ class DogsController < ApplicationController
 
   def destroy
     dog = Dog.find(params[:id])
+    unless current_user || dog.user.id == current_user.id
+      redirect_to root_path
+    end 
     dog.destroy
     redirect_to user_path(current_user.id)
   end
@@ -20,9 +23,6 @@ class DogsController < ApplicationController
 
   def dog_params
     dog_params_result=params.require(:dog).permit(:name,:avatar,:profile,:type).merge(user_id: current_user.id)
-    # unless dog_params_result[:avatar]
-    #   dog_params_result[:avatar] = File.read("app/assets/images/dog.jpeg")
-    # end
     return dog_params_result
   end
 end
